+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog for *TopoBank*
 
-## 1.53.0 (not yet released)
+## 1.53.0 (2025-01-23)
 
+- ENH: Route for querying pending or started analyses
 - BUG: Copy attachments and properties when deepcopying a DST
+- BUG: Return named results without specifying a name should not return all results
 
 ## 1.52.4 (2024-12-16)